SqlPython_Project/healthcare_etl_cli/requirements.txt
requests
psycopg2-binary
pandas
click  
